# Move length check to __init__ method of sequence