@marcelm 2021-12-13
One tiny issue is that we need to remember to check the length of sequence and qualities if we ever add a second call to `Sequence.__new__` in the future, but I think it’s worth it.